Naming and testing language model refactoring changes.
The change has been tested by comparing to an earlier KenLM
implementation, and with the same preprocessing, the same results
were observed on the tennis dataset (works on utterance-level).
Further, the cross-entropy refactoring has been tested using the
old ConvoKit surprise demo, and (with some randomness in choosing
the context and target) the outputs seem to be about the same,
i.e., almost the same ordering in most and least surprising
involvements. In the tennis demo, the values remained the same.